Possible fix for monster interaction.
The GameData.monsterChasing has been extended to short from byte so that it can keep track of which monster is chasing which player. The top byte will correspond to the monster chasing flag for the master client, while the lower byte for the slave client.

This fix should ensure that only the client whose player is being chased can stop the same monster from chasing. That is, if monster 1 is chasing master player, then only master client's request to stop the chase will be allowed by the server.

However, one issue that must be addressed for this to work correctly is to ensure missing packets are accounted for. Hence an acknowledge mechanism is required.

+++ core/src/com/mygdx/amaze/networking/GameData.java
@@ -20,7 +20,7 @@ public class GameData implements Serializable {
     public Coord friendPosition;
 
     // monster
-    public byte monsterChasing; // bit manipulation
+    public short monsterChasing; // bit manipulation
     public Coord[] monsterPosition;
 
     // item 
@@ -38,4 +38,3 @@ public class GameData implements Serializable {
     public byte requestType;
     public boolean requestOutcome;
 }
-
